fix - Update tax rate methods to return BigDecimal instead of double for consistency

+++ src/main/java/co/invest72/investment/domain/tax/StandardTax.java
@@ -29,7 +29,7 @@ public class StandardTax implements Taxable {
 	}
 
 	@Override
-	public double getTaxRate() {
-		return taxRate.getValue().doubleValue();
+	public BigDecimal getTaxRate() {
+		return taxRate.getValue();
 	}
 }
